+++ src/jax_fdm/goals/vertex/__init__.py
@@ -1,4 +1,11 @@
 from .vertex import *  # noqa F403
+from .point import *  # noqa F403
+from .coordinates import *  # noqa F403
+from .line import *  # noqa F403
+from .segment import *  # noqa F403
+from .plane import *  # noqa F403
+from .residual import *  # noqa F403
+from .colinear import *  # noqa F403
 from .normal import *  # noqa F403
 from .tangent import *  # noqa F403